requirements.txt
alembic==1.7.5
aniso8601==9.0.1
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
coverage==6.2
cryptography==36.0.0
dnspython==2.1.0
eventlet==0.33.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
Mako==1.1.6
MarkupSafe==2.0.1
outcome==1.1.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.27
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
zipp==3.6.0
